презентер для ивентов разделен на два для удобства работы с датабиндингом
