+++ src/main.js
@@ -1,8 +1,9 @@
 import HeaderPresenter from './presenter/header-presenter';
-import EventsPresenter from './presenter/events-presenter';
+// import EventsPresenter from './presenter/events-presenter';
+import TripPresenter from './presenter/trip-presenter';
 
 
 const bodyElement = document.querySelector(`.page-body`);
 
 new HeaderPresenter(bodyElement).init();
-new EventsPresenter(bodyElement).init();
+new TripPresenter(bodyElement).init();
