+++ src/presenter/trip-presenter.js
@@ -3,41 +3,34 @@ import DayList from "../view/day-list";
 import NoEvents from "../view/no-events";
 
 import TripDayItem from "../view/trip-day-item";
-import DayEvent from "../view/day-event";
-import EventEdit from "../view/event-edit";
+import EventPresenter from './event-presenter';
 
 import {
   render,
-  replace,
-  remove,
   RenderPosition
 } from "../utils/render.js";
 import {
   eventsArray,
-  renderEventModel
-} from "../utils/event-utils";
-import {
-  TRANSPORT_TYPE,
-  SERVICE_TYPE,
-  CITIES,
-  SortType
-} from "../const";
-import {
-  generatePhotoSrcs,
+  renderEventModel,
   sortTime,
   sortPrice,
   sortEvents
 } from "../utils/event-utils";
+import {
+  SortType
+} from "../const";
 
-export default class EventsPresenter {
+export default class TripPresenter {
   constructor(bodyContainer) {
     this._bodyContainer = bodyContainer;
     this._sortComponent = new Sort();
     this._dayListComponent = new DayList();
     this._noEventsComponent = new NoEvents();
     this._currentSortType = SortType.NO_SORT;
-    this._eventListeners = this._eventListeners.bind(this);
+    this._eventsPresenters = [];
+
     this._handleSortChange = this._handleSortChange.bind(this);
+
     this._eventModel = renderEventModel();
     this._initialEvents = this._eventModel.slice();
   }
@@ -91,58 +84,25 @@ export default class EventsPresenter {
     render(this._sortComponent.getSortContainer(), contentElement, RenderPosition.BEFOREEND);
 
     if (eventsArray.length > 0) {
-      this._getTripDayItems().forEach((item) => contentElement.appendChild(item.getElement()));
+      this._createTripDayItems().forEach((item) => contentElement.appendChild(item.getElement()));
     } else {
       render(contentElement, this._noEventsComponent, RenderPosition.BEFOREEND);
     }
   }
 
-  _getTripDayItems() {
+  _createTripDayItems() {
     return this._eventModel.map((model) => {
+
       const tripDayItem = new TripDayItem(model.date, model.index);
-      const dayEvents = this._createDayEvents(model.dayEvents);
-      const container = tripDayItem.getDayEventContainer();
-      dayEvents.forEach((event) => container.appendChild(event.getElement()));
+      const eventContainer = tripDayItem.getDayEventContainer();
+
+      model.dayEvents.forEach((event) => {
+        let eventPresenter = new EventPresenter(event, eventContainer);
+        this._eventsPresenters.push(eventPresenter);
+      })
+
+      this._eventsPresenters.forEach((eventPresenter) => eventPresenter.init());
       return tripDayItem;
     });
-  }
-
-  _createDayEvents(dayEvents) {
-    return dayEvents.map((event) => {
-      const dayEvent = new DayEvent(event);
-      const dayEditEvent = new EventEdit(event, TRANSPORT_TYPE, SERVICE_TYPE, CITIES, generatePhotoSrcs());
-      this._eventListeners(dayEvent, dayEditEvent);
-      return dayEvent;
-    });
-  }
-
-  _eventListeners(card, form) {
-    card.setClickHandler(() => {
-      replace(form, card);
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-    form.setEditSubmitHandler(() => {
-      replace(card, form);
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    });
-
-    form.setResetClickHandler(() => {
-      replace(card, form);
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    });
-
-    const onEscKeyDown = (e) => {
-      if (e.key === `Escape` || e.key === `Esc`) {
-        e.preventDefault();
-        replace(card, form);
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      }
-    };
-  }
-
-  _clearEventList() {
-    remove(this._dayListComponent);
-    this._getTripDayItems();
   }
 }
